requirements.txt
flask==3.0.3
gunicorn==22.0.0
authlib==1.3.2
requests==2.32.3
python-dotenv==1.0.1
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
psycopg2-binary>=2.9.9
